remove deprecated API from application component commit_hash:7e15909172b5be16683bd353db8be44f71332424

+++ client/android/div-histogram/src/main/java/com/yandex/div/histogram/NoOpHistogramBridge.kt
@@ -18,7 +18,7 @@ class NoOpHistogramBridge : HistogramBridge {
         min: Long,
         max: Long,
         unit: TimeUnit,
-        bucketCount: Long
+        bucketCount: Int
     ) = Unit
 
     override fun recordSparseSlowlyHistogram(name: String, sample: Int) = Unit
